build: require Go 1.27 and decode payloads with encoding/json/v2 only
Decision recorded in PERF_REFACTOR_PLAN.md: the minimum Go version moves
to 1.27.0 and the sonic decoder path is removed, leaving a single
standard-library engine. sonic drops back to an indirect dependency via
gin. CI pinned to 1.27.0.

- merge the json/v2 decoder into payload.go; drop payload_sonic.go and
  the go1.27 build tags; rename payload_jsonv2_test.go to
  payload_semantics_test.go.
- jsonv2Options still carries AllowDuplicateNames + AllowInvalidUTF8 so
  behaviour is unchanged for clients; switching to v2's strict defaults
  is a pending decision.
