+++ payload_semantics_test.go
@@ -1,5 +1,3 @@
-//go:build go1.27
-
 package cosy
 
 import (
